[UNTESTED] Add new "tube"-based DMA interface.
***NEW DMA INTERFACE***. Needs testing and experimentation in a
variety of common use cases before we commit to it.

STM32F1 conducts DMAed bytes through channels 1--7. STM32F2 conducts
them through streams, where stream listens to one of channels 0--7.
There is otherwise a close correspondence between channels and
streams: each is configured with a peripheral and memory address, a
number of data to transfer, and other settings (memory/peripheral
address post-increment, etc.). The STM32F2 versions are more
general (supporting double buffering and FIFOs, etc.), but otherwise
there is a direct analogy beween channels on F1 and streams on F2 (the
fact that F2 reused the word "channels" to describe what streams
listen to is confusing and unfortunate).

To help present a simple abstraction, we introduce the type
"dma_tube", which stands for whatever you configure to set up a DMA
transfer. On STM32F1, a dma_tube is a dma_channel. On STM32F2, it will
be a dma_stream. You can configure a dma_tube using the function
dma_tube_cfg(), which each series must implement.

This abstraction (along with some others, notably dma_request_src and
dma_mode_flags, described in comments in the patch to
<libmaple/dma.h>) lets us make a DMA interface that will hopefully
help portability between the two series. It's true that code using DMA
is, in the end, nonportable, but large chunks of it can be abstracted
out, with the nonportable bits hopefully mostly relegated to data
rather than code.

Implement this new abstraction on F1, keeping the old names (which use
"channel" instead of "tube") around for backwards compatibility. Also
update for XL-density and fix some DMA2 CH4/CH5 bugs we've had lying
around for a while. (This really belongs in another commit...).

Add <libmaple/dma_common.h> for things that <libmaple/dma.h> and
<series/dma.h> both need. This prevents us from putting things into
the libmaple header, then including the series header, which is an
ugly hack we've used elsewhere and should retroactively fix.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>