STM32F1 performance line: Rename __irq_adc -> __irq_adc_1_2.
This is the performance line ADC1/ADC2 IRQ handler. It shouldn't be
named __irq_adc, as that falsely indicates that it's the single global
IRQ handler for *all* ADCs (as is the case on STM32F2).

Note that on STM32F1 value line, there's only an __irq_adc1, as value
line MCUs only have a single ADC (and it's equivalent to ADC1 on
performance line). This will matter when you need to write IRQ
handlers for F1.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>